Reference puppet_gem better
pe_gem is supplanted by puppet_gem in Puppet 4.0. However, a user's
first inclination might be that puppet_gem is another module, and so
waste time looking for it on the Forge and getting confused. This commit
updates the README to explicitly state that it's a built-in component of
Puppet 4.0 and provide a direct link to the provider description in the
docs.